requirements.txt
numpy>=1.24.0
pillow>=9.0.0
gradio>=3.32.0 